Working on making sure the user is authorized
