+++ groovin-music/src/components/landingPage/landingPage.js
@@ -8,7 +8,6 @@ function LandingPage() {
   return (
     <>
       <body>
-        <NavBar />
         <div className="Header">
           <h1>GroovnMusic</h1>
         </div>
@@ -20,7 +19,7 @@ function LandingPage() {
           <input type="text" className="input" placeholder="Search ...." />
           <ul>...</ul>
         </div>
-        <Route exact path="/" component={LandingHome} />
+        <LandingHome />
       </body>
     </>
   );
